refactor: split into component files
